Update tests for pipeline run
